Create Analysis of Single Server Bulk Arrival Queueing System Using Simulation Method
The difference between the single server bulk arrival queueing system analysis simulation algorithm and the single server single arrival queueing system analysis simulation algorithm lies in the simulation stage for the arrival event. In the single server bulk arrival queueing system analysis simulation algorithm, data generation of the customer arrival group size is carried out first when starting the simulation for the arrival event which will function to update the number of customers in the system and the value of the length of the waiting line.